Fix overzealous filtering of unwanted class fields
Summary:
The definitions in a class scope should become class fields.

Before this diff, we handled type params (both new syntax and legacy) in the class scope itself. But this populates those names in the class scope, which then erroneously became fields.

D67886793 added a filter, which collects all names into a set before we build the class scope from the body's statements. It then skips those names when building the class fields. As pointed out in that diff, this causes us to filter out too much, since a class field might re-use a name.

This diff proposes a simple solution -- don't populate the class scope with type params in the first place. We create an extra annotation scope on class body entry. Type params (new syntax and legacy) are bound within that scope.

This changes a conformance test, but the change is (almost) a fix. We are actually finding the correct binding (the type param, not the global int), but the assert_type logic fails because internally it's a Quantified not a TypeVar.

Reviewed By: ndmitchell

Differential Revision: D69165600

fbshipit-source-id: cc057a2fb734163df07e6354755e107c3ae0d9dd