Get rid of pub(super)
Summary: Rust has a very rich way of expressing inter-dependencies. It's way too clever for its own good. Things that are private in a single file can have invariants ensured. As soon as you go beyond that, might as well make it pub and accept anyone can mess with it. Simpler.

Reviewed By: rchen152

Differential Revision: D69179263

fbshipit-source-id: 135043cfe9ab3485dc36418b9862a1d9f02066ca