Support namespace package in finder
Summary:
too_many_namespaces

In D69224139, we shifted Pyre2 to use the CPython module finding algorithm. But one important functionality missing there was namespace package. This diff fills in that gap and let us fully model the path-based import system.

Namespace packages were introduced in Python 3.3. Prior to that point, in order to import a module named `a.b`, CPython requires a file named `a/__init__.py` must exist on the filesystem. If that `__init__.py` file doesn't exist, CPython won't find `a.b` even if the file `a/b.py` also exists on the filesystem.

What namespace modules bring to the table was the possibility for CPython to treat directories without an `__init__.py` file to be a valid package as well. Another thing they did on top of it was to assign some special module finding behaviors to these non-`__init__.py`-backed modules: If multiple search roots all contain a directory with the same name, CPython would conceptually "merge" these directories, as opposed to having one search root "shadow" another one as they did with regular packages.

It's not too hard to extend our algorithm in D69224139 to encode such "merging" behavior: all we need to do is to (1) perform an additional is_directory check on the right paths after the `__init__` check, and (2) avoid the early return behavior like what we did when scanning for single-file modules or regular package.

Reviewed By: ndmitchell

Differential Revision: D69256720

fbshipit-source-id: 6969729ea56969398b28320ee2d798c0ff805a0c