implement Callable <: Protocol
Summary:
I think this has to be special-cased, since we check callable types against the type of the protocols `__call__` method. This diff covers assigning a callable-typed value to a protocol-typed variable, but not the inverse (assigning a protocol-typed value to a callable-typed variable)

Bound method representation can be refactored later so that the "drop first param" operation doesn't need to be repeated all over the place.

Reviewed By: ndmitchell

Differential Revision: D69193663

fbshipit-source-id: 48cc097931d9dd8c344302c8d2f75e3c05fe946e