Pass error collector on the call stack
Summary:
To support speculation, we want to be able to check code for errors without showing those errors to the user.

Per discussion on D69310187, we want to try passing the `&ErrorCollector` value through the call stack. For speculation, a caller can construct a fresh error collector and inspect it after the call.

This diff just removes the error collector from AnswerSolver and passes the error collector everywhere it's needed. This is a manual, but mechanical change. I did not try to change any APIs, just add the parameter if needed.

I left BindingsBuilder, another place we store the error collector in a struct, alone. We will not need to do any speculation in the bindings builder.

This work exposed a few awkward parts of the codebase.
1. We need a collector for is_subset_eq, but that doesn't make sense. This infects a bunch of code that just calls is_subset_eq, but otherwise doesn't need a collector (notably, `unions`).
2. `get_from_module` needs a collector, because `get` needs one. We should probably have separate collectors per module.

Reviewed By: ndmitchell

Differential Revision: D69337744

fbshipit-source-id: 1919fba2dd20434256009e8f4ce7804e42377edf