Add return expr key for None <: ann checks
Summary:
For NoReturn,  there are cases where we need to insert two different bindings for the same function (for example checking the annotation against none and checking the annotation against the return statement)

So I am making a new key specifically for the none checks. The none here is synthesized.

The main reason for the new key is so we do not clash with the preexisting ReturnExpression key.

I wasn't sure how to distinguish between when to use ReturnExpression vs the new key in this diff. It seems arbitirary? so I thought reserving the new key for none checks could be a way to distinguish them.

Reviewed By: ndmitchell

Differential Revision: D69342486

fbshipit-source-id: b9861c053f129f5bb8f1c9bc069559709ffafd09