Reset error collector at binding boundary
Summary:
D69337744 threaded the error collector through the call stack, but it went too far.

The binding boundary is unique. The computation that forces a binding's computation can vary. Requesting a computation will either force it or re-use an existing result. It does not make sense to use the context of the caller, including the error collector.

Instead, we should reset the error collector back to the "main" error collector for that module. This diff does this in a simple (perhaps simplistic) way: We re-introduce a field on the main AnswersSolver struct, pointing to the "base" collector for the module.

It's not correct to use this collector from arbitrary points -- code should almost always use the collector on the call stack. But we can use this field to reset when we go through bindings.

Reviewed By: ndmitchell

Differential Revision: D69381580

fbshipit-source-id: 1210f63aaaad84762842bfc0c4bf10d343ef976f