Make indexing by path work
Summary:
Causes us to use the Path in the handle. Which means we can have two different paths with the same module name in scope at once, in the state. We still ban this in check.

Note that we assume find is consistent. It will be, because we cache it, but better not to require that. It's a bit brittle. Will clean that up now.

Reviewed By: grievejia

Differential Revision: D69400515

fbshipit-source-id: b6355eae2d941ae9fa1ca5ace19d5d8295911be5