Add FlowStyle::FunctionDef
Summary:
Store the function idx in the flow environment, so we can chain functions with their predecessor definition.

It would be possible to look up the binding that corresponds to the flow key instead, but this was a bit awkward, since we need to call `self.table.types.1.get` and deal with optionals.

Using a FlowStyle is similar to what we already do with MergableImport, so it felt OK. Happy to change it.

Reviewed By: ndmitchell

Differential Revision: D69451319

fbshipit-source-id: 3058d4077fac3ba11601500d7c6015994cdf32b2