Add failing test for super() calls
Summary:
I honestly don't understand how the super() function is supposed to work. In a class method, it does not need arguments, but it can be called from the global scope with arguments as well.

I added the builtins.pyi definition from typeshed, so the  name is at least defined. We will certainly need to special case this somehow.

Reviewed By: ndmitchell

Differential Revision: D69611682

fbshipit-source-id: fab5361e465679356627d122c60fb7145f99ea5d