replace empty TypeVarTuple quantified & update quantified display
Summary:
Allow us to better differentiate the different kinds of quantified values in error messages.

The empty value for TypeVarTuple is a no-op, not sure if that code path is exercised. But I think this behavior should align with the spec: https://typing.readthedocs.io/en/latest/spec/generics.html#behavior-when-type-parameters-are-not-specified

> When a generic class parameterized by a type variable tuple is used without any type parameters and the TypeVarTuple has no default value, it behaves as if the type variable tuple was substituted with tuple[Any, ...]

Reviewed By: ndmitchell, migeed-z

Differential Revision: D69666775

fbshipit-source-id: 34aedcd290db8f93d6d026fd2ce6f20d593db33a