Remove unused Key::YieldTypeOfGenerator
Summary: We never look up the result of this key directly. If we don't care about the result, the only reason for a Key to exist is to ensure that the binding is solved and any errors are discovered. However, the binding itself is a Phi, which only stores `Idx`s, meaning the wrapped bindings will necessarily already be stored under different keys.

Reviewed By: ndmitchell

Differential Revision: D69704955

fbshipit-source-id: dcbea4cd277704287d2da9e2394a2e2e0ff6b749