Fix crash for yield outside a function
Summary:
During the answers phase, yield expressions always look up a key. Before this diff, we only added keys when binding functions. This led to crashes when we encountered a yield expression outside of a function, because we never added a key.

This diff fixes by ensuring that we always add a key. We already ensured that a FuncInfo value was in the bindings builder, and all invalid yield expressions are already accumulated there.

This diff adds a new "Invalid" binding for all of these yield expressions. With this change, the lookups in the answers phase will always succeed.

Reviewed By: ndmitchell

Differential Revision: D69730351

fbshipit-source-id: 6524a897ca0143ada3477af7c8a1543e95b68abd