Add failing tests for dict literal unpacking
Summary:
The unpacking logic has two bugs.

When there is a hint, we check the unpacked value against the hint directly. However, we can unpack a Mapping, which is not a subtype of dict (it's a supertype).

When there is no hint, we use decompose_dict to extract the key/value from the unpacked value. This is the wrong operation, and we should use unwrap_mapping instead. (Same issue as D69758244)

Reviewed By: ndmitchell

Differential Revision: D69764928

fbshipit-source-id: 6f92470ab98ad3c0f62190e25dea5eb28f9afe69