Add override check
Summary:
Our main question was "where should we implement the override check such that it happens exactly once?"

yangdanny97 suggested implementing the check  when we resolve `BindingClassField`, which has an advantage of having a range we can use to localize the error. Indeed, from the unittests, we can see that the error messages are at the field definition location.

The main goal of this diff is to determine where the checks should go. Once we have that, there are various cases to go over from there.

next steps:
- handle overloads
- handle when a member uses an override decorator but the method does not exist in the parent class. For this, how about we add a new field to BindingClassField for the decorators and then collect them at this point?https://www.internalfb.com/code/fbsource/[87d051ca273c]/fbcode/tools/pyre/pyre2/pyre2/bin/binding/class.rs?lines=126
- handle comparing properties like `x=3` against `x:int` which involves attribute subtyping handling those cases

Reviewed By: yangdanny97

Differential Revision: D69570486

fbshipit-source-id: df2f6c913477a3fbf199467af65b550d7551d5a4