Only consider call signature errors in overload selection
Summary:
This stack:

When we analyze a call, we also infer the types of the arguments. For overloaded calls, we try to find an overload where the call succeeds by checking if there are any errors.

This behavior is too strict. An argument expression might result in an error, but still evaluate to a type that is compatible with the function signature. These errors should not affect overload selection.

This diff:

When calling an overloaded signature, we create two separate error collectors: arg_errors and call_errors. We use the former to infer types of arguments, but we use the latter for arity issues, missing/duplicate arguments, and checking argument types against parameter types.

A signature matches if there are no call errors, even if there are arg errors.

Hacks:

The call_infer implementation too hacky, and we should clean it up soon. I spent a little while thinking about this, but gave up -- sorry! I've left inline comments on the hacks and invariants I assumed.

Reviewed By: ndmitchell

Differential Revision: D69901523

fbshipit-source-id: 27151ccdb6909ab45206de001a4ff5efa5f88946