Avoid some unnecessary usages of self.stdlib
Summary:
We had two ways of checking the identity of a class - by qualified name or by comparing it against the corresponding class in Stdlib. It's easy to cause bootstrapping panics when adding things to Stdlib, so it's better to standardize on checking by qualified name.

Also replaces a usage of stdlib.none_type with Type::None.

Reviewed By: stroxler

Differential Revision: D69956295

fbshipit-source-id: 49fc7cf98fe763f1cbf962700cb3aeb23e1050a6