Defer filtering ignored errors
Summary:
Whether an error is suppressed or not should not affect how that error affects overload selection, or any other speculative checking we want to do.

Instead of filtering ignored errors eagerly, we collect all errors and filter them out at the end. This is probably the right thing to do anyway, since we can give an option to print ignored errors as well.

Fixes the test in the diff stacked under this one.

Reviewed By: grievejia

Differential Revision: D69939819

fbshipit-source-id: 54e092726724d0978b057a28989e0849ebc793ed