Generalize class def <: type[T]
Summary:
Before this diff, we only supported type[Any] upper bounds, but other versions of this exist and are valid. Notably, we might want to instantiate a type var.

To fix, I needed to expose some more API through TypeOrder, specifically `promote_silently` to convert the Class value to a ClassType with default targs.

This is a case where it's difficult to error, since the type[] feature itself is designed this way, as far as I can tell. `type[C]` for a generic class `C` has unknown type args.

The type parameter default feature lets us instantiate with the default instead, but that's currently crashing pyre2, so I wasn't able to implement it in this diff.

Reviewed By: grievejia

Differential Revision: D69954717

fbshipit-source-id: 6b9d98460bd3962ef2961610ff97ce2c50e7f417