Determine if a class inherits from any
Summary:
I would like to determine if a parent class inherits from "Any" so I can allow overriding attributes that don't exist in the parent.

This is the plumbing for this flag.

Note that we found a bug/need further discussion on the "Any" type representation in pyre. Our typeshed is inconsistent with the real typeshed. One defines "Any" as object and another defines it as a class. This blocks me from properly adding the definition to stdlib and adding the class to the list of base classes. So instead, I use a flag to workaround the need to construct that type.

Reviewed By: stroxler

Differential Revision: D69954256

fbshipit-source-id: e529415b2014e9df5bbe69528a408c55798f1b4a