Add a test / TODO on Any resolution
Summary:
It turns out we are still using logic from when pyre2 was only a few weeks
old where we naively match an identifier against "Any" to decide whether it
is `typing.Any`.

This isn't right, we need to actually do the identity analysis. It's not yet clear
how to best handle this (we discovered that typeshed handling of `Any` is a little
weird which affects our options here), so for the moment I'm putting in a TODO rather
than trying to fix it right away; Jia is going to put a PR up to improve typeshed.

Reviewed By: rchen152

Differential Revision: D69954404

fbshipit-source-id: b2ba5c494c747531398a6b4610d57e85c10b2c90