Merge trailing negated ops in if statements
Summary:
The logic for if statements iterates over the syntactically apparent branches -- that is, if there is only an `if`, we will only run once.

But the logic for merging flow environments should always deal with the consequent and alternate branches -- even if the alternate branch is implicit.

In other words, the implicit alternate branch's flow environment contains the negated ops, just as an explicit one would have. We should merge them back into the joined flow regardless.

Reviewed By: rchen152

Differential Revision: D69971372

fbshipit-source-id: cd1388b0b4c617a8f2dca24bcfc3528e105f584d