Narrow elif test expressions
Summary:
Pretty straightforward. If we made it to the test expression, then the previous branch's predicates are negated.

I changed up the use_range associated with Key::Narrow. Each if/elif/else branch has its own range, so use that. For implicit else, use TextRange::default, which will be distinct from the others.

(I considered using Option<TextRange>, but this increases the size of Key.)

Reviewed By: ndmitchell

Differential Revision: D69971871

fbshipit-source-id: cfb9eeec21c8fdba27aabc7c212a05b3baba8c6c