Implement basic attribute go-to-definition support
Summary:
In this stack, we will implement basic support for go-to-definition on attributes.

This diff combines the ranges we started to track since the last diff with the `module_info` from the class type, and propagates this information all the way to `attr_infer` where we type check an attribute access. We then store the result in the traces for definitions.

There are a few cases where I have marked with TODO (e.g. when the base is a module), which we can add support for later.

In IDE services, we will look up the trace to find the relevant definition location for the attribute.

Reviewed By: stroxler

Differential Revision: D69948969

fbshipit-source-id: 9d1e180e0447fe27fd13d15bdd70519955a60c81