Re-implement attribute go-to-definition with `State::adhoc_solve`
Summary:
Before the last diff, we do not have the ability to call type checker code for LSP operations. As a result, go-to-definition for attributes is implemented in an awkward way: by recording every single definition result of attributes in a trace.

This diff now re-implements attribute go-to-definition with `State::adhoc_solve` introduced in the last diff. As a result, we no longer need to record all the definition locations in the trace.

Reviewed By: ndmitchell

Differential Revision: D70128413

fbshipit-source-id: 1db45b3a3ee7d8fd7521728bbcf4b8c3e04f5c6c