Expand the comment ClassFieldInitialValue more
Summary:
I got a bit confused about what was going on here and had to dig through
the plumbing a bit.

The existing comment was good, but I think tweaking it to emphasize two
different examples makes it clearer what's going on: assignment is a special
case where we can save an expr, all other cases we just know there is a value.

In addition, noting what's going on both in the type *and* in the `as_initial_value`
function makes it easier to connect the two.

Reviewed By: grievejia

Differential Revision: D70178420

fbshipit-source-id: ed6e016a63b8753308a7a443e67f003409c3e936