Add a test of dataclasses with methods
Summary:
I spent about a half hour trying to understand the dataclasses code, which looks
buggy in cases where it actually works because it relies on non-local interactions
between class metadata and class field logic.

In the procesds, I realized that there's a bug waiting to happen: we have an open
TODO to error on unannotated class fields, but the only reason we *don't* currently
have a bug in the handling of methods is because they are treated as unannotated.

If we tried to close the TODO without a unit test guarding against breakig method
handling, we'd potentially wind up with a huge regression on real codebases because
it's very common to define methods in dataclasses.

I'm putting a unit test of method handling in the test module, and placing it
directly above the TODO on `test_unannotated_attribute`

Reviewed By: yangdanny97, rchen152

Differential Revision: D70186956

fbshipit-source-id: 0832d732e79bf6b36514e88d3365faf5961924c0